Rename DB file
